Add ATSTK1000 and ATSTK1002 board support Patch by Haavard Skinnemoen, 06 Sep 2006
This patch adds support for the ATSTK1000 with the ATSTK1002 CPU
daughterboard.

ATSTK1000 is a full-featured development board for AT32AP CPUs. It
has two ethernet ports, a high quality QVGA LCD panel, a loudspeaker,
and connectors for USART, PS/2, VGA, USB, MMC/SD cards and
CompactFlash cards. For more information, please see this page:

http://www.atmel.com/dyn/products/tools.asp?family_id=682

The ATSTK1002 is a daughterboard for the ATSTK1000 supporting the
AT32AP7000 chip.

Signed-off-by: Haavard Skinnemoen <hskinnemoen@atmel.com>