Fixes bug clearing the bss section for i386
Hi,
There is a bug in the code of clearing the bss section for processor
i386.(File: cpu/i386/start.S)
In the code, bss_start addr (starting addr of bss section) is put into
the register %eax, but the code which clears the bss section refers to
the addr pointed by %edi.

This patch fixes this bug by putting bss_start into %edi register.

Signed-off-by: Mushtaq Khan <mushtaq_k@procsys.com>