Bugfix: remove embedded null (\0) from CFG_BOOTFILE macro in TQM8540_config
/bin/bash and /bin/dash (which /bin/sh is linked to on ubuntu) handle embedded
nulls in a string differently.  For example, the following statement:
    echo "this is a string\0" > afile
Will produce the following with /bin/bash:
    "this is a string\0"
But with /bin/dash, will produce:
    "this is a string

Bug fixed by moving the embedded null out of the makefile and into the
config header.  Also renamed the macro to avoid usage colision with the same
macro used by other board ports.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>