Zoom2 Fix serial gpmc setup
The offset to the chip select is incorrect.

The change 187af954cf7958c24efcf0fd62289bbdb4f1f24e,

omap3: embedd gpmc_cs into gpmc config struct

introduced a problem with the serial gpmc setup.

This patch reverts the chip select to its previous value.

The symptoms of this problem are that the Zoom2
currently hangs.

This was run tested on Zoom2.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>