x86: Fix do_go_exec() - const argv[]
Commit 54841ab50c20d6fa6c9cc3eb826989da3a22d934 made the argv parameter
to do_go_exec() const but did not allow for the fact that argv[-1] is
set to point to the global data structure and relies on argv being non-
const.

With this patch, do_go_exec() creates a new copy of the argv array with
an extra element to store global data pointer rather than simply
clobbering an arbitrary memory location.