arm/km: add boardid and hwkey to kernel command line
We need in some cases a possibility for the kernel to distinguish
on which board he is running. On powerpc we did this with different
dts files. On arm currently we can't do this, so add boardid and
hwkey to the kernel command line and use it later on in the kernel
code.

Signed-off-by: Holger Brunck <holger.brunck@keymile.com>