EFI PARTITION: Fix print_part_efi() and test_part_efi().
The functions suffered from passing parameters to_gpt_valid() and
is_pte_valid() incorrectly.  They should be passed a pointer to a
pointer that is written by the function.  Instead, they passing
the value of a pointer to a pointer.

In addition, neither of the functions was using the name of the
partition.  I added a simple function to generate an ASCII name
from the unicode partition name so that the name was usable.

I also cleaned up some compiler warnings, a little optimization,
and added error checking to ensure that accesses beyond the end
of the pgpt_pte array were not done,

Change-Id: I4ec22dbab24966c2ad8b919d239b95cdf914919d
Signed-off-by: Scott Anderson <saa@google.com>