PARTITION/BLOCK DEVICE: Add partition erase/read/write functions
To make accessing partitions easier and more reliable, this
change adds functions for finding, erasing, reading and writing
partitions.  A few notes:

1) Some partitions may need commands executed before and/or after
accessing them.  These functions will fetch an environment
variable of the form (pre|post)_<operation>.<partition_name> and
then hand it off to the command interpreter at the appropriate
time.  For example, if you have a partition "xloader" that
requires using hardware ECC, but all your other partitions need
software ECC, setting "pre_write.xloader=nandecc hw" and
"post_write.xloader=nandecc sw" will flip the ECC to hardware
before writing to the xloader partition and back to software when
finished.

2) The functions will silently round up the passed lengths to
meet the needs of the underlying block device.  This means that
if you call one of the functions to read or write and tell it to
only do one byte, you'll need to ensure that the passed buffer is
actually big enough to get an entire blocks worth of data.  In
the case of writing, if you want to prevent "random" data being
written after your one byte, you'll need to fill the buffer with
data of your choice.

Change-Id: Ibc4441c3c24a3f47549ae4b1ffad359a3d63a83d
Signed-off-by: Scott Anderson <saa@google.com>

Conflicts:

	disk/part.c
	include/part.h