smdk4412: enable to boot uImage with uncompressed kernel
'bootm' command checks ZIMAGE_MAGIC (0x16f2818) at offset 24h of the beginning
of the kernel image but failed with uImage since it has 64 bytes header.
This failure makes 'hdr' to be zero and failed to boot with invalid reference.

But this patch is also failed when uImage is loaded at the same address with
Entry Address. Therefore uImage should be loaded another address somewhere to
be copied itself to Load Address.

Signed-off-by: Dongjin Kim <tobetter@gmail.com>