smdk4412: Select FIMD bypass for LCD
The output of FIMD is connected to MIE/MDNIE as default on reset state,
therefore LCDBLK_CFG have to be configured to bypass FIMD of LBLK0.
It was configured by kernel in setup-fimd0.c, but nowhere to configure this
register in v3.10-rc1 since GPIO are configured by device tree and
setup-fimd0.c is removed.

Signed-off-by: Dongjin Kim <tobetter@gmail.com>