TI:armv7: Change CONFIG_SYS_SPL_ARGS_ADDR to a higher address
With changes to increase the size of the device tree (required to move
more data out of the kernel and into DT), loading the args at the old
address leads to us overwriting things later on.  To correct this, load
the args file to where we load the device tree anyhow.  This is also
safe for non-DT booting as in either case we use r2 to pass in the
location of things.

Signed-off-by: Tom Rini <trini@ti.com>