Allow use of fastboot on non-Android systems
This change allows fastboot to be used to flash boot partitions and
boot images even on a non-Android layout.  Rather than exit when a
non-android partitioning scheme is found when the user types
`fastboot' to the u-boot prompt, rely on checking that a partition
exists at flash or erase time.

Signed-off-by: Peter Chubb <peter.chubb@nicta.com.au>