FROMLIST: rockchip: make boot_mode related codes reused across all platforms
setup_boot_mode function use the same logic but different
mode register address across all the rockchip platforms,
so it's better to make this function reused across all the
platforms, and let the mode register address setting from
the config file.

Also add support for rk312x soc which is a little
special: the bootrom download flag is stored in
a grf register but the other boot mode flags are
stored in anohter pmugrf register.

Change-Id: I2e6a0ba870626adb837975c08094250d47767dac
Signed-off-by: Andy Yan <andy.yan@rock-chips.com>
Reviewed-by: Simon Glass <sjg@chromium.org>