board: rockchip: add rk3288-evb-rk1608 board support
The rk3288-evb-rk1608 is a development/evaluation for RK3288
and RK1608. The RK3288 as Host AP, and the RK1608 as a
Dual-cores DSP coprocessor, it provides high-performance
professional computing capabilities to Host AP.

They communicate with each other through MIPI and SPI. There
is a good host environment on host RK3288, we can easily debug
RK1608.

Also, the rk3288-evb-rk1608 is different with rk3288-evb, so
we need to add a independent BSP configuration for rk3288-evb-rk1608.

Change-Id: I75ec0e14ee88acd9de7f809d7e88bca97a77a5c2
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>