rockchip: resc_img: fix err if initialize resource list
In the function init_resource_list, if hdr is not NULL
The content addr is the sum of hdr addr and
(hdr->c_offset) * RK_BLK_SIZE. Then return 0 directly
without free hdr and content.

If hdr is NULL and gpt table is enabled, the parameter
offset must be multiplied by RK_BLK_SIZE and be divided
by RK_BLK_SIZE in function blk_read.

Change-Id: Ib4120745af8aa4924464f85c8bb4b34daee98fa6
Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>