UPSTREAM: usb: dwc2: make casts of ep->dma_buf consistent
In most places in the code we cast this to an unsigned long, but in one
place we cast to an unsigned int.  For consistency and to fix a warning
on 64bit targets, always cast this to unsigned long.  For the long term
we should however change the declaration of dma_buf.

Conflicts:
	drivers/usb/gadget/dwc2_udc_otg_xfer_dma.c

Change-Id: I56b7b6da2d716a9cd40ec4edbc928de433dfd4ca
Cc: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Acked-by: Lukasz Majewski <lukma@denx.de>
Acked-by: Marek Vasut <marex@denx.de>
Signed-off-by: Tom Rini <trini@konsulko.com>
(cherry picked from commit a4d403290e32f95cfd254f2778e08f655ddd5a49)