Import Android's dt_table.h for DT image format
PD#SWPL-772

Android documentation defines the recommended image format for storing
DTB/DTBO files in a single dtbo.img image. This patch includes the
latest header file with the struct definitions for this format from
AOSP.

The header was adapted to U-Boot's coding style and the function
declarations were removed.

Signed-off-by: Alex Deymo <deymo@google.com>
[trini: Change SDPX tag location]
Signed-off-by: Tom Rini <trini@konsulko.com>

Change-Id: Icb17cc449caf570169ddae4f3fd7e034ceef1734
Signed-off-by: jiamin ma <jiamin.ma@amlogic.com>