rockchip: dts: rv1108: correct the reference of u2phy
The reference of u2phy should be OTG port or host port in
the DT node of USB controller. Because the phy ops is bind
to each port or u2phy. So it will result in a NULL point if
we use the parent DT node as reference. This patch change
<&u2phy 1> to <&u2phy_otg> and <&u2phy_host> to solve this
problem.

Change-Id: I7e8fc5fcdeb1ce188c857bb3c5caa23aa713988e
Signed-off-by: MengDongyang <daniel.meng@rock-chips.com>