ODROID-C3/N2: add new fusing script
The script in this patch is to flash the bootloader image to a memory card,
the file 'u-boot.bin' in the current directory will be flashed by default.

    $ ./sd_fusing.sh <memory/card/device>

You can set the environment variable 'UBOOT' with a path of another bootloader
image file in another file or in another directory.

    $ UBOOT=<somewhere/in/your/machine> ./sd_fusing.sh <memory/card/device>

Change-Id: I9b4e14d1b3f0933a383eecaeb8726ee2dfdc40f1
Signed-off-by: Dongjin Kim <tobetter@gmail.com>