usb: gadget: rockusb: add vendor storage r/w cmd support
At present the vendor storage partition is small than 64K in the
design, so vendor storage r/w commands do not take multi-packet
into consideration, if it is needed, the codes should do a little
change in future. Besides, the RPMB r/w support need to be added
when its APIs get ready.

Change-Id: I621604c99b44c69eb3aff3e01b28a18f35303040
Signed-off-by: Frank Wang <frank.wang@rock-chips.com>