UPSTREAM: usb: dwc2: Add reset ctrl to driver
Add code to reset all reset signals as in usb DT node. A reset property
is an optional feature, so only print out a warning and do not fail if a
reset property is not present.

If a reset property is discovered, then use it to deassert, thus
bringing the IP out of reset.

Change-Id: I1c8560b20900d65d628c05df20f5e6e5555bc2cf
Signed-off-by: Ley Foon Tan <ley.foon.tan@intel.com>
Signed-off-by: Frank Wang <frank.wang@rock-chips.com>
(cherry picked from commit 88c34b8da62cfb4188ff9a59acb32f88ce9ed1b0)