UPSTREAM: usb: Don't enable CONFIG_SPL_DM_USB for non-DM SPL
When CONFIG_SPL_DM_USB was added, it was added defaulted on without
protection to ensure it didn't get set for non-DM SPL builds.

This leads to unexpected and confusing failures when building a DM based
U-Boot but with a non-DM SPL, as the defconfig is unlikely to have
CONFIG_SPL_DM_USB actively disabled.

Tweak the Kconfig to not enable this setting unless CONFIG_SPL_DM is set.

Change-Id: I11e59ec852e9621f01e526e21805a922ba1feb44
Signed-off-by: Martyn Welch <martyn.welch@collabora.com>
Signed-off-by: Frank Wang <frank.wang@rock-chips.com>
(cherry picked from commit 16ecb295b3834d8017d922ae46eedbbf6fd6c99b)