rockchip: boot mode: remove "recovery" partition lookup for recovery mode
Since we have unified android and linux image format that "recovery" partition
is necessary, we remove partition lookup when recovery mode is required.

Anyway, there is no recovery partition in A/B system and recovery ramdisk
has been moved to boot.img, and adding this judgment code is a little mess.

Change-Id: I9f590f464a14ac25855b8a9487c55cec8d353890
Signed-off-by: Joseph Chen <chenjh@rock-chips.com>