clk: rk3399: fix up the i2c clk error
I2c MUx is in cpll by default, but cpll is for dclk_vop exclusive.
If dclk_vop set rate after i2c init, the CPLL changed,
but the i2c not perception, it will resulting the wrong frequency
of the i2c.
So set the i2c frequency according to the kernel configuration.
and Hang I2C on the GPLL.

Change-Id: I91f891e9033e9d4648027ea253998a54011f4863
Signed-off-by: Elaine Zhang <zhangqing@rock-chips.com>