dm: serial: bind stdout device anyway
If the console is not marked to be bound, bind it anyway.

This makes serial node in U-Boot dtb no need to be enabled
and bound into dm tree since they are not sured to be the
stdout for pre-loader serial.

This avoids the mix of kernel and U-Boot serial node under
serial uclass list, otherwise serial device bound from U-Boot
dtb always have the higher priority.

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: I4d35626288124d14d47df2e7492d12f1af6a36ae