UPSTREAM: dm: core: Allow iterating devices without uclass_get()
At present we have uclass_foreach_dev() which requires that uclass_get()
be called beforehand to find the uclass. This is good if we suspect that
that function might fail, but often we know that the uclass is available.

Add a new helper which does this uclass_get() automatically, so that only
the uclass ID is needed.

Change-Id: I3106b199653bf8a32d2d1a0025b77261881716fc
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Bin Meng <bmeng.cn@gmail.com>
(cherry picked from commit 3cf0fba4ff862d833545f82fb2209ff3c79d17b5)
Signed-off-by: Cody Xie <cody.xie@rock-chips.com>