dts: Makefile: simplify the logic of kernel dtb feature
Our purpose is replacing the u-boot.dtb with u-boot-spl.dtb
when kernel dtb feature is enabled. The current logic is too
complicated and not easy to extend new features.

The best way is to generate but not replace a u-boot.dtb the
same as u-boot-spl.dtb during DTB compile process.

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: I678ec7c60c91ecc6f6f0a935e50a691c52c94162