BACKPORT: check-config: allow to complete build even with ad-hoc CONFIG
Currently, the check-config.sh terminates the build when unknown
ad-hoc options are detected.  I think it is too much because we may
want to patch config headers locally in a build/deployment project.

So, let's relax check-config.sh to just warn even if it detects
options that are not in the whitelist.  Instead, this check can be
done at the end of build, along with other checks.  It will catch
more attention.

Even with this change, the Buildman tool catches new warnings,
so Tom can give NACK to new ad-hoc options.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Dongjin Kim <tobetter@gmail.com>
Change-Id: I0c068de48a58b65b8bdec7579ee10042f9c48dc4