common: fdt_support: remove earlycon if uart is disabled
If uart is required to be disabled during power on, it would be
not initialized by any pre-loader and U-Boot.

If we don't remove earlycon from commandline, kernel hangs while
using earlycon to putc/getc which may dead loop for waiting uart
status. (It seems the root cause is baundrate is not initilalized)

So let's remove earlycon from commandline.

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: I2edcf48cb4dd301f2e79d5e94b1477c55198708f