UPSTREAM: pinctrl: Return -ENOSYS when system call is not available
Update the code to use -ENOSYS, which is the correct error code for an
unimplemented system call in U-Boot.

Also we should not check for a missing operations array as this is not
permitted. For now this can be covered by an assert().

Change-Id: Ia554867bd25ece4ceb3214325768ac59dab1f66f
Signed-off-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit bddac45d0422acbba5bf01406d27eedb429a1f42)
Signed-off-by: Cody Xie <cody.xie@rock-chips.com>