phy: Add set_mode callback
The initial use for this is for PHYs that have a mode related to USB
OTG. There are several SoCs (e.g. TI OMAP and DA8xx) that have a mode
setting in the USB PHY to override OTG VBUS and ID signals.

Of course, the enum can be expaned in the future to include modes for
other types of PHYs as well.

Change-Id: Ic8b5f33f928341a88132cb203b65f156f75d18b6
Signed-off-by: Wyon Bi <bivvy.bi@rock-chips.com>