mmc: change the clock firstly in hs400 mode
In U-Boot, if change MMC device to hs400, it has to change to hs200
firstly. Then change to hs ddr mode. This may occur error that change
the mode firstly since the clock is higher than the hs ddr mode needed.
So change the clock firstly before change to hs ddr mode.

Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>
Change-Id: Ic564579cbd7dad9376a27b2102c83c802c702dbd