cpu: rockchip amp: add cpu boot-on control
If there are subnode under "/rockchip-amp/amp-cpus"
in kernel fdt, it means they are expected to be boot
by linux, so U-Boot don't boot it.

Otherwise we can add "boot-on = <..>" value in amp.its to
control boot it or not. If this prop is absent, it's treat
as "boot-on = <1>" by default.

Priority: kernel fdt > amp.its.

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: Ic0a86ab86140ea8a8546c670ab07bcca7d128c59