drivers: power: add pd chip driver
This temporary patch is to solve the problem that the rk3588s
tablet can be turned on when the battery is out of power.
This Type-C PD driver only supports PD fast charging function,
compatible with PD fusb302 and husb311 chip. If DFP/UFP detection
and dp altmode functions are added later, we will revert this patch.

Signed-off-by: Wang Jie <dave.wang@rock-chips.com>
Change-Id: Ic3411725ef6d9bcc4ca48b272ef67c4c9975077f