usb: dwc3: gadget: start new transfer on XFER_COMPLETE
In current code, the next transfer in dwc3 queue needs to wait for a
following XFERNOTREADY event coming after the current transfer is
completed. However, the gadget in U-Boot awlays uses polling method
to check the DWC3 interrupt, so the gadget class driver is required
to poll dwc3 interrupt twice to check the XFERCOMPLETE and XFERNOTREADY
event separately in ping-pong buffer usage scenario.

So let start the new transfer in dwc3 request list right away when the
time we get to XFERCOMPLETE event for non-sync transfer, instead of
waiting for the following XFER_NOT_READY event.

Change-Id: I113c6551606acd4fc0d0fbafb84d066179eec587
Signed-off-by: Frank Wang <frank.wang@rock-chips.com>