rockchip: Introduce CONFIG_ROCKCHIP_BOOTDEV
Usually we get boot device from preloader or scan list on the
board with single storage. While for the multiple storage, we
introduce this option to determine which we really want to be
the boot device, which contains kernel, rootfs and etc.

When this option is NULL string, we fall through to get boot
device from preloader or scan list.

Example: CONFIG_ROCKCHIP_BOOTDEV="nvme 0".

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: I734dc0ab2bb7104a1584cbddc15620c890970223