rockchip: rk3288: Locate fdt at 0x08100000
Before this patch, the FDT(Flattened Device Tree) blob is located at
0x08300000, and the following 13MB from 0x08400000 is used as OPTEE
memory which will be set to be secure.

The kernel will map at least 2MB (4MB if LPAE=y) for the FDT blob, make
the 0x08400000 be mapped to the kernel, cpu or device access them will fail.

When LPAE is enabled, 0x08100000 is round down to 0x08000000 and then
map 4MB size which is not over 0x08400000, so 0x08100000 is also safe.

Such as
[  106.766128] 8<--- cut here ---
[  106.769224] Unhandled fault: imprecise external abort (0x406) at 0xb5fde000
[  106.776209] pgd = 94a412ca
[  106.778936] [b5fde000] *pgd=043c3835, *pte=703a175f, *ppte=703a1c7f

Change-Id: Ia4efb0dad695996e3f825805d4c6e336a7e4d1a3
Signed-off-by: Jianqun Xu <jay.xu@rock-chips.com>