usb: dwc3: gadget: fix dcache invalidate for event buffer
The second parameter of invalidate_dcache_range() is the end address
to invalidate up to not the invalidation length, so wrap a new API
dwc3_invalidate_cache() make the driver easy to use.

Move up the invalidation operation into event buffer check instead of
event buffer process to reduce the invoked times.

Fixes: 0e7e455ba5bd ("usb: dwc3: invalidate dcache before read event buffer")
Change-Id: I299740e5d15e1f1841d5b33cd1cd2f9e34ebcc12
Signed-off-by: Frank Wang <frank.wang@rock-chips.com>