dm: core: pinctrl: Select state "init" for pinctrl driver
We'd like to let pinmux stay in a expected state during boot,
even if the user driver is not implemented in u-boot
but in kernel.

Let the pinctrl driver to select the init state of pins,
and the user driver can select them to default state
in u-boot or later in kernel driver probe.

Change-Id: Ie16d0f01afdfab1a4e5be168c350cd661cb6408c
Signed-off-by: Cody Xie <cody.xie@rock-chips.com>