usb: gadget: rockusb: avoid enter maskrom if usb cable not connected
On platforms which support rockusb3.0, when we use "reboot loader"
to restart system to loader mode and USB cable is not connected,
in this case USB3 polling will timeout and USB speed will switch
to USB2 and re-enumerate. But on current code, -ENODEV will be
returned which cause entering rockusb failed and error entering
maskrom.

error log:
   RKUSB: LUN 0, dev 0, hwpart 0, sector 0x0, count 0x3a3e000
   wait for usb3 connect timeout
   RKUSB: LUN 0, dev 0, hwpart 0, sector 0x0, count 0x3a3e000
   Enter rockusb failed, fallback to bootrom...

Signed-off-by: Jianwei Zheng <jianwei.zheng@rock-chips.com>
Change-Id: I3ecd408ea1605ad81839274f167353f3de9eb6c8